+++ src/niconamaClient.js
@@ -1,7 +1,7 @@
 // @flow
-import User from './user';
-import Live from './live';
-import LiveCommentStream from './liveCommentStream';
+import User from './user/user';
+import Live from './live/live';
+import LiveCommentStream from './live/commentStream/liveCommentStream';
 
 export default class NiconamaClient{
   liveClient: Live;
@@ -16,32 +16,18 @@ export default class NiconamaClient{
     this.liveClient = new Live(liveId, session);
   }
 
-  createCommentStream(): LiveCommentStream {
+  createLiveCommentStream(): LiveCommentStream {
     if (!this.liveClient) throw new Error('Live instance is not defined');
     return this.liveClient.getCommentStream();
   }
 
   liveComments(): Promise<any> {
-    if (!this.liveClient) throw new Error('Live instance is not defined');
+    if (!this.liveClient) Promise.reject('Live instance is not defined');
     return this.liveClient.comments();
-
-    //return this.liveClient.rpResult()
-    //  .then(xmlString => { return this.liveClient.comments(xmlString) })
-    //  .catch( err => {
-    //    return Promise.reject(err);
-    //  })
   }
 
-  //doLiveComment(liveId: string, session: string, comment: string) {
-  //  return this.liveClient.rpResult(liveId, session)
-  //    .then(xmlString => { return this.liveClient.liveInfo(xmlString) })
-  //    .then( data => { return this.liveClient.testDoComment(comment) })
-  //    .then( data => {
-  //      return Promise.resolve(data);
-  //    })
-  //    .catch( err => {
-  //      return Promise.reject(err);
-  //    })
-  //    ;
-  //}
+  doLiveComment(comment: string): Promise<any> {
+    if (!this.liveClient) Promise.reject('Live instance is not defined');
+    return this.liveClient.doComment(comment);
+  }
 }
